sesion16/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pymongo